Add "Use IdP user ID" login method.
In some cases, when user domains can be mixed, and users can change their email address at the remote provider, we don't have a reliable attribute to relate local accounts to remote accounts. This introduces the use of the user ID from the remote IdP stored as a user meta on the local WP install and used to identify the logged in user. The user meta item is keyed as `{idp_name}_idp_user_id`.